Msim: Initialize subscriptions once on info change
- When the subscription info for active subs changes,
we iterate through the list of active subscriptions
info list. If either the display name or the subscription
id of a SIM card changes we immediately initialize the
subscription list with the new information
- This causes a ConcurrentModificationException
as we read and write to the same list
- Check if any of the SIM info has changed
when we are iterating through the list. Initialize
the subscriptions only after we are done checking
the information

Change-Id: I56a34c4ebfb328395bb92ce49ac9c066e0fcbaab
CRs-Fixed: 906538